refactor: add redux folder
