+++ src/redux/mode/mode.actions.js
@@ -1,0 +1,17 @@
+import { OPEN_SEARCH, CHANGE_SORTING } from './mode.constants';
+
+export const openSearch = () => (
+  {
+    type: OPEN_SEARCH,
+    payload: {}
+  }
+);
+
+export const changeItem = newCriterion =>(
+  {
+    type: CHANGE_SORTING,
+    payload: {
+      chosenParameter: newCriterion
+    }
+  }
+);
