+++ src/redux/movies/movies.constants.js
@@ -1,0 +1,6 @@
+export const START_SEARCH = 'START_SEARCH';
+export const OPEN_SEARCH = 'OPEN_SEARCH';
+export const OPEN_MOVIE = 'OPEN_MOVIE';
+export const CHANGE_SORTING = 'CHANGE_SORTING';
+export const RESULTS_TO_STORE = 'RESULTS_TO_STORE';
+export const RECOMMENDED_TO_STORE = 'RECOMMENDED_TO_STORE';
