+++ src/redux/search/search.actions.js
@@ -1,0 +1,12 @@
+import { START_SEARCH } from './search.constants';
+
+export const startSearch = (phrase, type) => (
+  { 
+    type: START_SEARCH,
+    payload: {
+      searchPhrase: phrase,
+      searchType: type,
+      replaceCurrentResultsBy: [],
+      loadingData: true
+    }
+});
