+++ src/redux/search/search.constants.js
@@ -1,0 +1,1 @@
+export const START_SEARCH = 'START_SEARCH';
